@@ -10,10 +10,10 @@
         }
     ],
     "require": {
-        "php": "~7.0",
-        "phpunit/phpunit": ">= 6.0",
+        "php": "~7.1",
+        "phpunit/phpunit": ">= 7.0",
         "mikey179/vfsStream": "~1.6",
-        "squizlabs/php_codesniffer": "~3.1"
+        "squizlabs/php_codesniffer": "~3.2"
     },
     "suggest": {
         "codeception/codeception": "BDD testing framework for testing applications",
